Extract runner test helpers
